Cleanup search classes and serializers
