--- app/serializers/search/username_serializer.rb
@@ -1,8 +1,0 @@
-module Search
-  class UsernameSerializer < ApplicationSerializer
-    attributes :id,
-               :name,
-               :profile_image_90,
-               :username
-  end
-end
