+++ spec/initializers/rack/attack_spec.rb
@@ -11,7 +11,7 @@ describe Rack::Attack, type: :request, throttle: true do
   describe "search_throttle" do
     it "throttles /search endpoints based on IP" do
       Timecop.freeze do
-        allow(Search::Postgres::Username).to receive(:search_documents).and_return({})
+        allow(Search::Username).to receive(:search_documents).and_return({})
 
         valid_responses = Array.new(5).map do
           get "/search/usernames", headers: { "HTTP_FASTLY_CLIENT_IP" => "5.6.7.8" }
