+++ spec/services/search/article_spec.rb
@@ -2,7 +2,7 @@ require "rails_helper"
 
 # rubocop:disable Rails/PluckId
 # This spec uses `pluck` on an array of hashes, but Rubocop can't tell the difference.
-RSpec.describe Search::Postgres::Article, type: :service do
+RSpec.describe Search::Article, type: :service do
   describe "::search_documents" do
     it "returns an empty result if there are no articles" do
       expect(described_class.search_documents).to be_empty
